Better signature for Mapping.get().
Fixes #278.

The approach here is different from what we thought before.  According
to this commit, there are three valid forms of Mapping[K, V].get():

  d.get()      # returns Optional[V]
  d.get(None)  # also returns Optional[V]  (seen in the wild)
  d.get(v)     # requires v to have type V, and returns V

So we don't allow d.get(d) where d is a type unrelated to V.  This is
a compromise, but I don't think there's much use for d.get(d) with a
default that is differently-typed than V, and when I tried it my
application code (that didn't use this) wouldn't type-check correctly.