Allow asyncio.wait to accept non-Task Futures
asyncio.wait accepts Futures which are not Tasks.  I had some code
like this:

    my_future = asyncio.Future()

    # ...

    @asyncio.coroutine
    def my_coro():
        # ...

        yield from asyncio.wait([my_future], timeout=TIMEOUT)

Prior to this change, mypy complains about the first argument to
wait being a List[Future] rather than a List[Task].  After this
change, either is acceptable, which is consistent with the asyncio
documentation and code.