Change scandir() to type as both iterator and context-manager
Summary:
The existing typing for scandir() assumes it can only be an iterator. It can also be used as a context manager per the documentation, and typing fails when used as part of mypy-0.530.

Hat-tip to a colleague, M. Pieters, for the guidance on how to do this.

Test Plan:
Wrote a small snippet of code that implements the context manager usage.
Ran stock mypy against that code, got warnings about exit and enter.
Ran amended mypy (ie, with this typing included) against that code, got no warnings.