Add a tox.ini to make running all tests easier
I've been using this for a while now and it solves a couple of problems
for me.  Firstly, I was never running the pytype tests because it's
Python 2.7 which introduces manual switching when I'm primarily testing
with mypy; tox automates the creation and switching of the virtualenv.
Secondly, I was sometimes pushing broken code because of local changes
to mypy which was used when I was running the tests; tox creates a
pristine virtualenv using upstream mypy, which avoids this.