Base concurrent.futures.Executor on ContextManager
Update the concurrent.futures so that Executor, ThreadPoolExecutor, and ProcessPoolExecutor inherit from the typing.ContextManager[T] generic introduced in Python 3.6.

Specifically, this solves a problem in code such as:

with ThreadPoolExecutor() as p: ...

Where the p variable would be typed as an abstract `Executor`, rather than the specific `ThreadPoolExecutor` as expected.