Change the type of BaseException.message to Text.
This is debatable: an ASCII-only unicode object is acceptable, but
non-ASCII characters will cause the message to be replaced with
'<exception str() failed>'.  However I found some internal code that
sometimes assigns unicode objects to .message, and I'd rather not
burden that code with .encode() calls -- especially since in Python 3
this is a text string.  Note that if you pass the unicode object to
the constructor it type-checks, and ditto if you assign to .args.