Get rid of BLOCKED label in favor of [WIP] subject prefix (#1956)
We (I) introduced the BLOCKED label ~2 years ago, but it never got much usage, and I remember it startled some developers (Ivan may recall...). A better convention seems to put [WIP] in the subject, so let's standardize on that and get rid of the BLOCKED label.

I already removed the BLOCKED label and edited the subject of the one issue under it.

We also don't make good use of the other labels, but they seem more useful and less controversial.