add io.TextIOWrapper.read in 3.3
This is needed to make TextIOWrapper and a few classes that inherit
from it concrete in 3.3. I am actually not sure whether this method
exists at runtime; there's nothing in
https://docs.python.org/3/library/io.html#io.TextIOWrapper
that suggests it was added in 3.4. In any case, it hardly matters
since 3.3 usage should be very rare now.

This appears to be the last typeshed change needed to unblock
python/mypy#5223. @ilevkivskyi would you mind reviewing?