Loosen parameters of format_list and from_list
Never ask for a List when asking for a MutableSequence will do. Never
ask for a MutableSequence when asking for a Sequence will do. Never ask
for a Sequence when asking for a Collection will do. Never ask for a
Collection when asking for an Iterable will do.

Also drop support for Tuple[str, int, str, Optional[str]] elements in
the iterables passed to traceback.format_list and
traceback.StackSummary.from_list - if authors are diligent enough to be
type-checking their code, they are diligent enough to be making use of
their dependencies according to their preferred semantics rather than
their backwards-compatibility carve-outs.