Make urllib2's OpenerDirector.open not optional
This is a follow-up to #2821, though maybe more controversial.  It
seems like it is more plausible that a directly instantiated opener
could actually fail to handle a request, though I'm not familiar
enough with use patterns to be sure.

This fixes a small handful of errors in a dropbox code base.
I could definitely see saying no on this one. Thoughts?