sys: refine the return type of sys.exc_info()
I believe that either all of the items are None or none of them are.
Reflect that in the type.

This was attempted initially in
https://github.com/python/typeshed/pull/246
but apparently mypy didn't allow unpacking given such a union, but it
does now (checked with mypy 0.711).

The (incorrect) TODO from that PR was removed in commit
25ac4d6af46104d0c7cb134ed09a61ca7092b62a but the commit message noted
that this change is desirable.