Update __import__ function annotations
Per the docs, globals/locals is an optional argument.

https://docs.python.org/3/library/functions.html#__import__
https://docs.python.org/2/library/functions.html#__import__

Additionally, globals/locals can be any mapping type, not only a dict.

Likewise, fromlist can be any sequence (the docs mention a tuple, not a
list).

The function returns a ModuleType, not Any.

Now matches the definition in stdlib/3/importlib/__init__.pyi.