Remove some spurious version checks in logging
PR #3269 added some version checks for the argument type to setLevel
and the existence of NullHandler. While these features weren't present
in early versions of Python 3, they *are* present in Python 2.7, which
leads to false positives.

We could expand the check to include Python 2, but since the affected
Python 3 versions are super old and unsupported, I'm just removing the
checks.